pyproject.toml
[project]
name = "z-auth-and-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "fastapi-mail>=1.4.2",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
